Improve DNS lookup performance.
DNS lookups where serialized in the agent. A bad request therefore
blocked good requests for some time. Also, a lookup was performed in
the traffic-manager even though an agent had responded. An agent
response should be authoritative even if it didn't find anything.

This commit ensures that the agent produces lookup responses in a
separate goroutine to ensure that multiple lookups can run in parallel
and ensures that no lookup is made by the traffic-manager if responses
have arrived from at least one agent.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>